requirements.txt
aws-cdk-lib==2.182.0
aws-cdk.aws-apigatewayv2-alpha==2.114.1a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.114.1a0
aws-cdk.aws-lambda-python-alpha==2.182.0a0
constructs>=10.0.0,<11.0.0
